Reorganize block-validation directory
The two entry points: `block-validation` and `block-editor` are both used in the context of the block editor. Potentially they could share the same entry point so that code re-usability would be better.

This is the first step of the process. The directory structure of the `block-validation` is now more organized and should match the structure of the `block-editor` folder.

This change will also allow us to cleanly migrate the `amp-toggle` plugin from `block-editor` to `block-validation`. With #5997, this plugin logically belongs to `block-validation` since it renders AMP status notification that in turn is based on the AMP validation status.
