+++ assets/src/block-validation/components/error/error-content.js
@@ -19,8 +19,8 @@ import { BlockIcon } from '@wordpress/block-editor';
 /**
  * Internal dependencies
  */
-import AMPAlert from '../../../images/amp-alert.svg';
-import AMPDelete from '../../../images/amp-delete.svg';
+import AMPAlert from '../../../../images/amp-alert.svg';
+import AMPDelete from '../../../../images/amp-delete.svg';
 import { getErrorSourceTitle } from './get-error-source-title';
 
 /**
