+++ assets/src/block-validation/components/error/error-panel-title.js
