+++ assets/src/block-validation/components/error/error-type-icon.js
@@ -12,9 +12,9 @@ import {
 /**
  * Internal dependencies
  */
-import HTMLErrorIcon from '../../../images/amp-html-error-icon.svg';
-import JSErrorIcon from '../../../images/amp-js-error-icon.svg';
-import CSSErrorIcon from '../../../images/amp-css-error-icon.svg';
+import HTMLErrorIcon from '../../../../images/amp-html-error-icon.svg';
+import JSErrorIcon from '../../../../images/amp-js-error-icon.svg';
+import CSSErrorIcon from '../../../../images/amp-css-error-icon.svg';
 
 /**
  * Component rendering an icon representing JS, CSS, or HTML.
