+++ assets/src/block-validation/components/icon/index.js
@@ -6,9 +6,10 @@ import PropTypes from 'prop-types';
 /**
  * Internal dependencies
  */
-import AMPLogoIcon from '../../images/amp-logo-icon.svg';
-import AMPToolbarIcon from '../../images/amp-icon-toolbar.svg';
-import AMPToolbarIconBroken from '../../images/amp-toolbar-icon-broken.svg';
+import './style.css';
+import AMPLogoIcon from '../../../../images/amp-logo-icon.svg';
+import AMPToolbarIcon from '../../../../images/amp-icon-toolbar.svg';
+import AMPToolbarIconBroken from '../../../../images/amp-toolbar-icon-broken.svg';
 
 /**
  * Plugin icon.
