+++ assets/src/block-validation/components/sidebar-notification/index.js
